requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
ultralytics>=8.0.0
pyyaml>=6.0.0
